NBTDocument#write() can now be called without arguments. This feature is intended to make reading, modifying, and then writing changes back to the same NBT document more convenient. NBTDocuments returned by jnbt.read() will automatically set "write arguments" (target and compression) on the NBTDocuments they return, except in cases where these cannot be automatically determined (e.g. when source is a file-like object with no "name" attribute, such as io.BytesIO()). These write arguments will then be automatically provided to write() when called without arguments.
If write arguments are not set on an NBTDocument, then calling write() without arguments will raise a TypeError.
jnbt.read() allows you to provide a write target manually via the "target" parameter.
In addition, you can also set and get write arguments manually on an NBTDocument with the following methods:
    * NBTDocument#setWriteArguments()
    * NBTDocument#getWriteArguments()

The "compression" parameter is now ignored when reading/writing NBT files with jnbt.read()/NBTDocument#write() if the source/target is a file-like object (as opposed to the name of a file).

Fixed a bug in NBTDocument#__repr__ that cut off a few characters at the beginning of the contents string.
Fixed a bug in WrongTagError#__str__ that was preventing a description of the exception from being printed properly.

jnbt is now version 1.0.3.