jnbt.read() has a new optional parameter, create. If the given source is a str and the source file cannot be found, create determines how jnbt.read() handles the situation. If create is True, a new blank NBTDocument is returned. The new document will not be written to disk until .write() is called. If create is not True, then FileNotFoundError will be raised. create defaults to False.
jnbt.TAG_Compound objects have several new methods:
* TAG_Compound#setdefault_byte( name, ... )
* TAG_Compound#setdefault_bytearray( name, ... )
* TAG_Compound#setdefault_compound( name, ... )
* TAG_Compound#setdefault_double( name, ... )
* TAG_Compound#setdefault_float( name, ... )
* TAG_Compound#setdefault_int( name, ... )
* TAG_Compound#setdefault_intarray( name, ... )
* TAG_Compound#setdefault_list( name, ... )
* TAG_Compound#setdefault_long( name, ... )
* TAG_Compound#setdefault_short( name, ... )
* TAG_Compound#setdefault_string( name, ... )

These methods work like python's dict#setdefault( key, default ), but with the convenience of TAG_Compound#byte( name, ... ), etc.
These methods check if the TAG_Compound contains a tag with the given name.
If there is a tag with the given name and the type matches (e.g. TAG_Compound#setdefault_byte expects the existing tag to be a TAG_Byte), then the existing tag is returned.
If there is a tag with the given name but types do not match, a jnbt.WrongTagError is raised.
If there is not a tag with the given name, a new tag is created and the
additional arguments / keyword arguments passed to the setdefault_* method are passed to the constructor of the tag. The new tag is then inserted into the TAG_Compound and then returned.

jnbt is now version 1.0.11.