jnbt can now read Region (.mcr) worlds! To this end, the old code in world.py was rewritten and split across several files. The new code is pretty rough at the moment, but this will be refined over the course of several updates to come. Minecraft-specific code is now a part of the jnbt.mc package. World-related code is now a part of the jnbt.mc.world package.
Added jnbt.getWorld( name, savedir ) and jnbt.iterWorlds( savedir ) as convenience methods for opening and iterating over worlds.
In both cases, savedir is an optional path that the worlds will be looked for in that defaults to <your minecraft directory>/saves.

The Player class is now more versatile. A uuid is no longer required; it may be initialized with a name, in which case the UUID will be looked up from the Mojang API similarly to how name lookups with UUIDs work now.
Likewise, the nbt parameter is optional now as well. You can initialize a Player with a path to a gzipped NBT file, in which case nbt will be loaded from that path when it is first accessed.

World#iterPlayers() now searches both playerdata/ and players/ (for compatibility with older worlds). It also has two new parameters with the same name (both defaulting to True) that control whether or not to search the respective directory if it exists.

World#getBlockName() now correctly returns None if no name for the given id could be located in the block data dictionary for that world.

World, Dimension, Region, Chunk, and Block now have .format and .formatid class attributes. format is a string like "anvil" or "region", and formatid is a corresponding numerical code (e.g. 4 or 3)

Updated the README to advertise the new Region compatibility, show new usages introduced in this and the previous update, and add new resource links.

jnbt is now version 1.0.13.