Python 3.7 compatibility. Certain generators in JNBT were previously failing due to their reliance on StopIteration exception behavior; this behavior was changed in Python 3.7 (see PEP 380).
Fixed a bug with TAG_List tag appender/inserter methods.
If appending/inserting a tag in an empty list, if the tag constructor raised an exception the list's tag type would be changed but the tag would not be inserted.

Fixed a bug with TAG_List.__iadd__ (+= operator).
It was allowing you to add tags of any type to a list that stored tags of a different type.

Changed _rget_leaf() (implements rget() for TAG_String, TAG_Int_Array, TAG_Byte_Array).
Providing more than one positional argument now returns default instead of raising an exception.

jnbt is now version 1.0.14.